+++ package.json
@@ -15,6 +15,7 @@
     "axios": "^1.6.2",
     "modern-normalize": "^2.0.0",
     "react": "^18.2.0",
+    "react-datepicker": "^4.24.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.49.2",
     "react-redux": "^9.0.4",
@@ -22,8 +23,8 @@
     "react-router-dom": "^6.21.0",
     "react-select": "^5.8.0",
     "react-toastify": "^9.1.3",
-    "redux-persist": "^6.0.0",
     "react-transition-group": "^4.4.5",
+    "redux-persist": "^6.0.0",
     "styled-components": "^6.1.1"
   },
   "devDependencies": {
