+++ src/pages/SignInPage/SignInPage.jsx
@@ -1,31 +1,7 @@
-// import React from "react";
-// import { useForm } from "react-hook-form";
-
-import {
-  StyledWrap,
-  StyledHead,
-  StyledInput,
-  StyledSignInBtn,
-  StyledLink,
-  StyledFormWrap,
-} from "./SignInPage.styled.js";
+import SignIn from '../../components/SignIn/SignIn.jsx';
 
 const SignInPage = () => {
-  // const { register } = useForm();
-
-  return (
-    <StyledWrap>
-      <StyledFormWrap>
-        <StyledHead>Sign In</StyledHead>
-
-        <StyledInput name="email" placeholder="Email" />
-        <StyledInput name="password" placeholder="Password" />
-        <StyledSignInBtn>Sign In</StyledSignInBtn>
-
-        <StyledLink to="/signup">Sign Up</StyledLink>
-      </StyledFormWrap>
-    </StyledWrap>
-  );
+  return <SignIn />;
 };
 
 export default SignInPage;
